+++ app/pb_kits/playbook/pb_time_stacked/time_stacked.rb
@@ -15,6 +15,9 @@ module Playbook
                   default: false
       prop :classnames, type: Playbook::Props::String,
                         default: nil
+      prop :tag, type: Playbook::Props::Enum,
+                 values: %w[body caption],
+                 default: "body"
 
       def classname
         generate_classname("pb_time_stacked_kit", dark_class, classnames)
@@ -40,6 +43,10 @@ module Playbook
 
       def format_timezone
         pb_date_time.to_timezone
+      end
+
+      def tag_classnames
+        "pb_time_stacked_kit_#{tag}"
       end
 
     private